Separate out getAnagramSet.
